# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateQueueRequest:
    properties:
      closes_at:
        type: string
      lesson_id:
        type: string
      max_size:
        type: integer
      opens_at:
        type: string
      subject_id:
        type: string
    required:
    - opens_at
    - subject_id
    type: object
  dto.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorDetail'
      success:
        type: boolean
    type: object
  dto.GroupResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.PaginationResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
  dto.QueueListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.QueueResponse'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationResponse'
    type: object
...
  dto.QueueResponse:
    properties:
      closes_at:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserMinimalResponse'
      group:
        $ref: '#/definitions/dto.GroupResponse'
      id:
        type: string
      lesson_id:
        type: string
      max_size:
        type: integer
      opens_at:
        type: string
      slots_count:
        type: integer
      status:
        type: string
      subject:
        $ref: '#/definitions/dto.SubjectResponse'
    type: object
...
  dto.SubjectResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  dto.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
  dto.UpdateQueueRequest:
    properties:
      closes_at:
        type: string
      max_size:
        type: integer
      status:
        type: string
    type: object
...
  dto.UserMinimalResponse:
    properties:
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API for Stud Hub backend.
  title: Stud Hub API
  version: "1.0"
paths:
...
  /queues:
    get:
      description: Returns queues filtered by group and status with pagination.
      parameters:
      - description: Group ID (UUID). If omitted, current user's group is used.
        in: query
        name: group_id
        type: string
      - description: Queue status filter
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.QueueListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: List queues
      tags:
      - Queues
    post:
      consumes:
      - application/json
      description: Creates a queue for the authenticated user's group.
      parameters:
      - description: Create queue request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateQueueRequest'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Create queue
      tags:
      - Queues
  /queues/{queue_id}:
    delete:
      description: Deletes a queue by ID.
      parameters:
      - description: Queue ID
        in: path
        name: queue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Delete queue
      tags:
      - Queues
    get:
      description: Returns detailed queue information by queue ID.
      parameters:
      - description: Queue ID
        in: path
        name: queue_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.QueueResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Get queue by ID
      tags:
      - Queues
    patch:
      consumes:
      - application/json
      description: Updates queue status and limits.
      parameters:
      - description: Queue ID
        in: path
        name: queue_id
        required: true
        type: string
      - description: Update queue request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateQueueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.QueueResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      - CookieAuth: []
      summary: Update queue
      tags:
      - Queues
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"